Remove StackdriverStatsExporter.registerView(). Rearrange tests.
