+++ exporters/stats/stackdriver/src/main/java/io/opencensus/exporter/stats/stackdriver/StackdriverExporterWorkerThread.java
@@ -16,8 +16,11 @@
 
 package io.opencensus.exporter.stats.stackdriver;
 
+import com.google.api.MetricDescriptor;
 import com.google.cloud.monitoring.v3.MetricServiceClient;
+import com.google.common.annotations.VisibleForTesting;
 import com.google.common.collect.Lists;
+import com.google.monitoring.v3.CreateMetricDescriptorRequest;
 import com.google.monitoring.v3.CreateTimeSeriesRequest;
 import io.opencensus.common.Duration;
 import io.opencensus.stats.View;
@@ -26,7 +29,6 @@ import io.opencensus.stats.ViewManager;
 import java.util.HashMap;
 import java.util.List;
 import java.util.Map;
-import javax.annotation.concurrent.GuardedBy;
 
 /**
  * Worker {@code Thread} that polls ViewData from Stats library and batch export to StackDriver.
@@ -35,16 +37,10 @@ import javax.annotation.concurrent.GuardedBy;
  */
 final class StackdriverExporterWorkerThread extends Thread {
 
-  private static final Object monitor = new Object();
-
   private final long scheduleDelayMillis;
   private final String projectId;
   private final MetricServiceClient metricServiceClient;
-
-  @GuardedBy("monitor")
   private final ViewManager viewManager;
-
-  @GuardedBy("monitor")
   private final Map<View.Name, View> registeredViews = new HashMap<View.Name, View>();
 
   StackdriverExporterWorkerThread(
@@ -60,31 +56,44 @@ final class StackdriverExporterWorkerThread extends Thread {
     setName("ExportWorkerThread");
   }
 
+  // Registers the given view to Stackdriver Monitoring if the view is not registered yet, otherwise
+  // does nothing. Throws IllegalArgumentException if the given view is a different view that has a
+  // same name with a registered view.
+  @VisibleForTesting
   void registerView(View view) {
-    synchronized (monitor) {
-      View existing = registeredViews.get(view.getName());
-      if (existing != null) {
-        if (existing.equals(view)) {
-          // Ignore views that are already registered.
-          return;
-        } else {
-          throw new IllegalArgumentException(
-              "A different view with the same name is already registered: " + existing);
-        }
+    View existing = registeredViews.get(view.getName());
+    if (existing != null) {
+      if (existing.equals(view)) {
+        // Ignore views that are already registered.
+        return;
+      } else {
+        // TODO(songya): Do we still need to check this? Or can we assume that
+        // ViewManager.getAllExportedViews() will always return correct results?
+        throw new IllegalArgumentException(
+            "A different view with the same name is already registered: " + existing);
       }
-      registeredViews.put(view.getName(), view);
-      viewManager.registerView(view);
+    }
+    registeredViews.put(view.getName(), view);
+
+    // TODO(songya): don't need to create MetricDescriptor for RpcViewConstants once we defined
+    // canonical metrics. Registration is required only for custom view definitions. Canonical
+    // views should be pre-registered.
+    MetricDescriptor metricDescriptor =
+        StackdriverExportUtils.createMetricDescriptor(view, projectId);
+    if (metricDescriptor != null) {
+      metricServiceClient.createMetricDescriptor(
+          CreateMetricDescriptorRequest.newBuilder().setMetricDescriptor(metricDescriptor).build());
     }
   }
 
-  // Polls ViewData from Stats library for all registered views, and upload them as TimeSeries to
+  // Polls ViewData from Stats library for all exported views, and upload them as TimeSeries to
   // StackDriver.
-  private void export() {
+  @VisibleForTesting
+  void export() {
     List<ViewData> viewDataList = Lists.newArrayList();
-    synchronized (monitor) {
-      for (View.Name view : registeredViews.keySet()) {
-        viewDataList.add(viewManager.getView(view));
-      }
+    for (View view : viewManager.getAllExportedViews()) {
+      registerView(view);
+      viewDataList.add(viewManager.getView(view.getName()));
     }
     CreateTimeSeriesRequest.Builder builder = CreateTimeSeriesRequest.newBuilder();
     for (ViewData viewData : viewDataList) {
