+++ exporters/stats/stackdriver/src/test/java/io/opencensus/exporter/stats/stackdriver/StackdriverStatsExporterTest.java
@@ -16,49 +16,14 @@
 
 package io.opencensus.exporter.stats.stackdriver;
 
-import static org.mockito.Matchers.any;
-import static org.mockito.Matchers.eq;
-import static org.mockito.Mockito.atLeast;
-import static org.mockito.Mockito.doNothing;
-import static org.mockito.Mockito.doReturn;
-import static org.mockito.Mockito.timeout;
-import static org.mockito.Mockito.times;
-import static org.mockito.Mockito.verify;
-
-import com.google.api.MetricDescriptor;
-import com.google.api.gax.rpc.UnaryCallable;
 import com.google.auth.oauth2.GoogleCredentials;
-import com.google.cloud.monitoring.v3.MetricServiceClient;
-import com.google.cloud.monitoring.v3.stub.MetricServiceStub;
-import com.google.common.collect.ImmutableMap;
-import com.google.monitoring.v3.CreateMetricDescriptorRequest;
-import com.google.monitoring.v3.CreateTimeSeriesRequest;
-import com.google.monitoring.v3.TimeSeries;
-import com.google.protobuf.Empty;
 import io.opencensus.common.Duration;
-import io.opencensus.common.Timestamp;
-import io.opencensus.stats.Aggregation.Sum;
-import io.opencensus.stats.AggregationData.SumDataLong;
-import io.opencensus.stats.Measure.MeasureLong;
-import io.opencensus.stats.View;
-import io.opencensus.stats.View.AggregationWindow.Cumulative;
-import io.opencensus.stats.View.Name;
-import io.opencensus.stats.ViewData;
-import io.opencensus.stats.ViewData.AggregationWindowData.CumulativeData;
-import io.opencensus.stats.ViewManager;
-import io.opencensus.tags.TagKey;
-import io.opencensus.tags.TagValue;
 import java.io.IOException;
-import java.util.Arrays;
-import java.util.List;
-import org.junit.Before;
 import org.junit.Rule;
 import org.junit.Test;
 import org.junit.rules.ExpectedException;
 import org.junit.runner.RunWith;
 import org.junit.runners.JUnit4;
-import org.mockito.Mock;
-import org.mockito.MockitoAnnotations;
 
 /** Unit tests for {@link StackdriverStatsExporter}. */
 @RunWith(JUnit4.class)
@@ -66,41 +31,8 @@ public class StackdriverStatsExporterTest {
 
   private static final String PROJECT_ID = "projectId";
   private static final Duration ONE_SECOND = Duration.create(1, 0);
-  private static final TagKey KEY = TagKey.create("KEY");
-  private static final TagValue VALUE = TagValue.create("VALUE");
-  private static final String MEASURE_NAME = "my measurement";
-  private static final String MEASURE_UNIT = "us";
-  private static final String MEASURE_DESCRIPTION = "measure description";
-  private static final MeasureLong MEASURE =
-      MeasureLong.create(MEASURE_NAME, MEASURE_DESCRIPTION, MEASURE_UNIT);
-  private static final Name VIEW_NAME = Name.create("my view");
-  private static final String VIEW_DESCRIPTION = "view description";
-  private static final Cumulative CUMULATIVE = Cumulative.create();
-  private static final Sum SUM = Sum.create();
-
-  @Mock private ViewManager mockViewManager;
-
-  @Mock private MetricServiceStub mockStub;
-
-  @Mock
-  private UnaryCallable<CreateMetricDescriptorRequest, MetricDescriptor>
-      mockCreateMetricDescriptorCallable;
-
-  @Mock private UnaryCallable<CreateTimeSeriesRequest, Empty> mockCreateTimeSeriesCallable;
 
   @Rule public final ExpectedException thrown = ExpectedException.none();
-
-  @Before
-  public void setUp() {
-    MockitoAnnotations.initMocks(this);
-
-    doReturn(mockCreateMetricDescriptorCallable).when(mockStub).createMetricDescriptorCallable();
-    doReturn(mockCreateTimeSeriesCallable).when(mockStub).createTimeSeriesCallable();
-    doReturn(null)
-        .when(mockCreateMetricDescriptorCallable)
-        .call(any(CreateMetricDescriptorRequest.class));
-    doReturn(null).when(mockCreateTimeSeriesCallable).call(any(CreateTimeSeriesRequest.class));
-  }
 
   @Test
   public void createWithNullCredentials() throws IOException {
@@ -134,7 +66,7 @@ public class StackdriverStatsExporterTest {
   }
 
   @Test
-  public void createHandlerTwice() throws IOException {
+  public void createExporterTwice() throws IOException {
     StackdriverStatsExporter.createWithCredentialsAndProjectId(
         GoogleCredentials.newBuilder().build(), PROJECT_ID, ONE_SECOND);
     try {
@@ -144,94 +76,6 @@ public class StackdriverStatsExporterTest {
           GoogleCredentials.newBuilder().build(), PROJECT_ID, ONE_SECOND);
     } finally {
       StackdriverStatsExporter.unsafeSetExporter(null);
-    }
-  }
-
-  @Test
-  public void registerViewAndExport() throws IOException {
-    View view =
-        View.create(VIEW_NAME, VIEW_DESCRIPTION, MEASURE, SUM, Arrays.asList(KEY), CUMULATIVE);
-    ViewData viewData =
-        ViewData.create(
-            view,
-            ImmutableMap.of(Arrays.asList(VALUE), SumDataLong.create(1)),
-            CumulativeData.create(Timestamp.fromMillis(100), Timestamp.fromMillis(200)));
-    doNothing().when(mockViewManager).registerView(view);
-    doReturn(viewData).when(mockViewManager).getView(VIEW_NAME);
-
-    try {
-      StackdriverStatsExporter.unsafeSetExporter(
-          new StackdriverStatsExporter(
-              PROJECT_ID, new FakeMetricServiceClient(mockStub), ONE_SECOND, mockViewManager));
-      StackdriverStatsExporter.registerView(view);
-
-      verify(mockStub, times(1)).createMetricDescriptorCallable();
-      // createTimeSeriesCallable() should be called once per second. This test uses a timeout that
-      // is longer than the export interval in order to decrease the chance of failures due to
-      // timing.
-      // TODO(songya): consider how to avoid blocking the thread and making this test deterministic
-      verify(mockStub, timeout(2000).atLeast(1)).createTimeSeriesCallable();
-
-      MetricDescriptor descriptor = StackdriverExportUtils.createMetricDescriptor(view, PROJECT_ID);
-      List<TimeSeries> timeSeries =
-          StackdriverExportUtils.createTimeSeriesList(viewData, PROJECT_ID);
-      verify(mockCreateMetricDescriptorCallable, times(1))
-          .call(
-              eq(
-                  CreateMetricDescriptorRequest.newBuilder()
-                      .setMetricDescriptor(descriptor)
-                      .build()));
-      verify(mockCreateTimeSeriesCallable, atLeast(1))
-          .call(eq(CreateTimeSeriesRequest.newBuilder().addAllTimeSeries(timeSeries).build()));
-    } finally {
-      StackdriverStatsExporter.unsafeSetExporter(null);
-    }
-  }
-
-  @Test
-  public void preventRegisterViewBeforeCreateExporter() throws IOException {
-    View view1 =
-        View.create(VIEW_NAME, VIEW_DESCRIPTION, MEASURE, SUM, Arrays.asList(KEY), CUMULATIVE);
-    thrown.expect(IllegalStateException.class);
-    thrown.expectMessage("Stackdriver stats exporter has not been created.");
-    StackdriverStatsExporter.registerView(view1);
-  }
-
-  @Test
-  public void preventRegisteringDifferentViewWithSameName() throws IOException {
-    doNothing().when(mockViewManager).registerView(any(View.class));
-    StackdriverStatsExporter.unsafeSetExporter(
-        new StackdriverStatsExporter(
-            PROJECT_ID, new FakeMetricServiceClient(mockStub), ONE_SECOND, mockViewManager));
-    View view1 =
-        View.create(VIEW_NAME, VIEW_DESCRIPTION, MEASURE, SUM, Arrays.asList(KEY), CUMULATIVE);
-    StackdriverStatsExporter.registerView(view1);
-    View view2 =
-        View.create(
-            VIEW_NAME,
-            "This is a different description.",
-            MEASURE,
-            SUM,
-            Arrays.asList(KEY),
-            CUMULATIVE);
-    try {
-      thrown.expect(IllegalArgumentException.class);
-      thrown.expectMessage("A different view with the same name is already registered: ");
-      StackdriverStatsExporter.registerView(view2);
-    } finally {
-      StackdriverStatsExporter.unsafeSetExporter(null);
-    }
-  }
-
-  /*
-   * MetricServiceClient.createMetricDescriptor() and MetricServiceClient.createTimeSeries() are
-   * final methods and cannot be mocked. We have to use a mock MetricServiceStub in order to verify
-   * the output.
-   */
-  private static final class FakeMetricServiceClient extends MetricServiceClient {
-
-    protected FakeMetricServiceClient(MetricServiceStub stub) {
-      super(stub);
     }
   }
 }
